features/cve: rework the `RewriteStatement` prompt
commit dbf7e88496490b9c25aed312ecfe22f7e5587da3 brought in a proposal
that had not been approved by Product Security and it still raises many
questions that have not been answered until today.  Replace this draft
by the official guidance for writing a Statement that is effective
today.

Cover the changes by evals so that we do not change the behavior until
we decide to do so.

Related: https://issues.redhat.com/browse/AEGIS-53
PR: https://github.com/RedHatProductSecurity/aegis/pull/70