fix: deduplicate the definition of Tekton pipelines
The whole `pipelineSpec` subtree (as provided by the Konflux generator)
is identical for the `pull-request` and `push` pipelines.  It does not
make any sense to change everything in two separate files that are
*almost* identical with each other.  This commit moves the pipeline
definition to a separate file, which is referenced using the git
resolver.

Related: https://issues.redhat.com/browse/AEGIS-118