fix: simplify `AegisFeatureModel` to reduce LLM overhead
The descriptions of `confidence`, `completeness`, and `consistency` are
extremely complicated.  The full descriptions are being sent as input
tokens on each LLM prompt but the values assigned by LLM to these
metrics do not correlate with correctness of the suggestion anyway.
Moreover  `completeness` and `consistency` are not presented to OSIM
users at all.  So we are wasting LLM bandwidth for little to no benefit.

This removes the `completeness` and `consistency` fields and simplifies
the description of `confidence`.  Thanks to these changes, the average
duration per suggestion is reduced by 30% and the number of consumed LLM
tokens is reduced by 50% in some cases.  No undesired consequences are
observed.

Related: https://issues.redhat.com/browse/AEGIS-206
Fixes: https://github.com/RedHatProductSecurity/aegis-ai/issues/290