evals: tweak the versioning info detection rubric
Before this change, the `suggest-description-for-CVE-2025-54770`
evaluation case was still occasionally failing with:
```
The suggested_title contains "GRUB2", which includes versioning information ("2"), and this is not part of an acronym explanation within the title. The rubric states that versioning info is not allowed except in acronyms explanations.
```

Fixes: commit 1e70aed17d3ee3e927604a68a95304825b33626e
Related: https://issues.redhat.com/browse/AEGIS-264