enhance: monitor long running LLM prompts
If we processed multiple LLM prompts in parallel, it was not obvious
from the log which of them started to hang.  This change improves the
observability in such cases by logging an INFO message each 60s for
each LLM prompt that is still running.

Related: https://issues.redhat.com/browse/AEGIS-273